Move Logger::Levels enum to top level header
