+++ source/common/common/BUILD
@@ -23,6 +23,7 @@ envoy_cc_library(
     hdrs = ["assert.h"],
     deps = [
         ":minimal_logger_lib",
+        "//envoy/common:logger",
         "@abseil-cpp//absl/base",
         "@abseil-cpp//absl/debugging:stacktrace",
         "@abseil-cpp//absl/debugging:symbolize",
@@ -242,6 +243,7 @@ envoy_cc_library(
         ":lock_guard_lib",
         ":macros",
         ":non_copyable",
+        "//envoy/common:logger",
         "//source/common/protobuf",
         # Depend on the header-only interface, not the implementation lib, so that every commit to
         # main does not invalidate the build cache of minimal_logger_lib (and everything that
